Update cmb2-conditionals.php
Use CMB2 hooks instead of native WordPress hooks. No longer need to determine if CMB2 is loaded yet which resolves cases where CMB2 is loaded as part of a plugin that loads after this one. No longer to determine which page to load scripts on by piggy backing on CMB2 own logic.